confirm hapusdata, menambahkan edit data

+++ pages/datamahasiswa.php
@@ -47,7 +47,9 @@ require 'function.php';
             <td><?= $mhs["nim"] ?></td>
             <td><?= $mhs["jurusan"] ?></td>
             <td><?= $mhs["nohp"] ?></td>
-            <td><a href="hapusdata.php?id=<?= $mhs["id"] ?>"><button>Hapus</button></a></td>
+            <td><a href="hapusdata.php?id=<?= $mhs["id"] ?>"onclick="return confirm(Yakin?');"><button style="margin-bottom:12px;background-color blue;">Hapus</button></a></td>
+
+            <td><a href="ubahdata.php?id=<?= $mhs["id"] ?>"><button style="margin-bottom:12px;background-color: antiquewhite;">Edit</button></a></td>
             
             
         </tr>
